refactor: re-organize types and errors
