+++ src/errors/PrismicError.ts
