+++ src/lib/castArray.ts
@@ -8,4 +8,6 @@
  *
  * @returns `a` as an array.
  */
-export const castArray = <A>(a: A | A[]): A[] => (Array.isArray(a) ? a : [a]);
+export const castArray = <A>(a: A | A[]): A[] => {
+	return Array.isArray(a) ? a : [a];
+};
