+++ src/lib/castThunk.ts
@@ -7,5 +7,6 @@
  *
  * @returns `a` as a a thunk.
  */
-export const castThunk = <A>(a: A | (() => A)): (() => A) =>
-	typeof a === "function" ? (a as () => A) : () => a;
+export const castThunk = <A>(a: A | (() => A)): (() => A) => {
+	return typeof a === "function" ? (a as () => A) : () => a;
+};
