+++ test/client-dangerouslyGetAll.test.ts
@@ -6,7 +6,19 @@ import { mockPrismicRestAPIV2 } from "./__testutils__/mockPrismicRestAPIV2";
 import { testAbortableMethod } from "./__testutils__/testAbortableMethod";
 import { testGetAllMethod } from "./__testutils__/testAnyGetMethod";
 
-import { GET_ALL_QUERY_DELAY } from "../src/client";
+/**
+ * The number of milliseconds in which a multi-page `getAll` (e.g. `getAll`,
+ * `getAllByType`, `getAllByTag`) will wait between individual page requests.
+ *
+ * This is done to ensure API performance is sustainable and reduces the chance
+ * of a failed API request due to overloading.
+ *
+ * IMPORTANT: This value is linked to `GET_ALL_QUERY_DELAY` used in
+ * `../src/createClient.ts`. The two values do not need to be kept in sync, but
+ * note that changing `GET_ALL_QUERY_DELAY` in the public code may have an
+ * effect on tests using this test-specific constant.
+ */
+export const GET_ALL_QUERY_DELAY = 500;
 
 /**
  * Tolerance in number of milliseconds for the duration of a simulated network
@@ -131,7 +143,7 @@ it("throttles requests past first page", async (ctx) => {
 	const endTime = Date.now();
 
 	const totalTime = endTime - startTime;
-	const minTime = numPages * queryDelay + (numPages - 1) * GET_ALL_QUERY_DELAY;
+	const minTime = numPages * queryDelay + (numPages - 1) * 500;
 	const maxTime = minTime + NETWORK_REQUEST_DURATION_TOLERANCE;
 
 	// The total time should be the amount of time it takes to resolve all
